Some optimisations

+++ src/main/java/tgcfs/EA/Mutation/RandomResetting.java
@@ -7,6 +7,7 @@ import tgcfs.NN.EvolvableModel;
 
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.IntStream;
 
 /**
@@ -68,7 +69,7 @@ public class RandomResetting extends Individual {
      * @param model model to assign to the individual
      * @param myInputandOutput input output last
      */
-    public RandomResetting(INDArray objPar, Integer fitness, EvolvableModel model, List<TrainReal> myInputandOutput){
+    public RandomResetting(INDArray objPar, AtomicInteger fitness, EvolvableModel model, List<TrainReal> myInputandOutput){
         super(objPar, fitness, model, myInputandOutput);
     }
 
@@ -101,6 +102,6 @@ public class RandomResetting extends Individual {
      * @return RandomResetting object
      */
     public RandomResetting deepCopy(){
-        return new RandomResetting(this.getObjectiveParameters(), new Integer(this.getFitness()), this.getModel().deepCopy(), this.getMyInputandOutput());
+        return new RandomResetting(this.getObjectiveParameters(), new AtomicInteger(this.getFitness().intValue()), this.getModel().deepCopy(), this.getMyInputandOutput());
     }
 }
